Upgrade plotting to proper x-axis control
Now the plot shows proper X axis value based on the control, and uses
spd_raw for values. Thus, no interpolated data anymore.

Also tweaks X-Axis control a bit, with a todo to fix the ClampedSpinbox
behavior. Because it's super annoying to edit max now, when min is
non-trivial value.