OO: Improve automatic exposure mode (somewhat)
Now it converges quickly on a reasonable integration time, and tries
to circumvent the Flame-S acquisition "feature" (free running mode)
that normally returns bogus value after changes to integration time.

See
https://github.com/ap--/python-seabreeze/issues/110#issuecomment-3478107206

The root cause for that behavior, as far as I understand it, is that
Flame-S free-runs the entire time (keeps acquiring spectra in the
background). Thus by the time change to integration time comes,
there's usually another capture in progress, using the old parameter.

And weirdly enough, that capture is returned back, no worries.

This is problematic when switching between integration times when
trying to find an optimal integration time. Hence, waits (for old
value of IT) between changes to IT become necessary.

Possible improvement would be to put a façade in front of the set
integration time that consistently blocks for the old int-time.