Add float_to_string func that maximizes precision
This is useful because existing `e`, `f`, `g` format specs on their own
are rather... sucky; when you have a fixed buffer and want to utilize it
to the max:

``` sh
$ python -q
>>> a = 1.234567891e12
>>> "%.8g" % a
'1.2345679e+12'
>>> "%.9g" % a
'1.23456789e+12'
>>> "%.8g" % -a
'-1.2345679e+12'
```

And yes, in this case, using `%.9e` for positive and `%.8e` for negative
is the right call. But for other numbers (e.g. smaller ones) it might
not be. So `float_to_string` is a glorified wrapper around the logic
that tries to maximize the buffer space.